Bingo Vender
The “Bingo Office Supplies” store  sells up to 20 products, the details of which it keeps in the file “Products.txt” .  
This file has one line records consisting of  
•	the product code (5 characters) 
•	product name of up to 20 characters 
•	the number in stock  
•	the product cost price  
•	the name of the file holding an image of the product.  (at most 20 characters) 

These fields are separated by commas. 
When the company makes a sale of a product the following information needs to be recorded on a Sales Invoice. The company makes a maximum of 20 sales per day.

•	The Sales Invoice number (Start at any number and increment by 1) 
•	The product name
•	The product selling price (mark-up of 60%) 
•	The quantity purchased.
•	The value of the sale